requirements.txt
openai~=1.59.5
python-dotenv~=1.0.1
numpy~=1.26.4
scipy~=1.11.4
scikit-learn~=1.4.2
pandas~=2.2.3
openpyxl~=3.1.5
matplotlib~=3.7.5